+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed issue where there would be extremely visible lines in terrain textures.
 - Fixed issue with ADTs failing to export due to bad/missing doodads and/or WMOs.
 - The 'Open Export Directory' link after exporting has been replaced with 'View in Explorer' which opens the directory of the last exported item.
+- Fixed issue that prevented doodad sets from exporting/previewing under certain circumstances.
 
 0.1.8 (11-01-2020)
 - NPC variant textures can now be selected/exported for creature models.