+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed issue with ADTs failing to export due to bad/missing doodads and/or WMOs.
 - The 'Open Export Directory' link after exporting has been replaced with 'View in Explorer' which opens the directory of the last exported item.
 - Fixed issue that prevented doodad sets from exporting/previewing under certain circumstances.
+- 3D model panning via W, A, S, D, Q, E keys implemented.
 
 0.1.8 (11-01-2020)
 - NPC variant textures can now be selected/exported for creature models.